move most of the ugly stuff to lib/server.js
