+++ server.js
@@ -1,29 +1,10 @@
-const express = require('express')
-const promisify = require('pify')
 const chalk = require('chalk')
-const app = express()
+const createServer = require('./lib/server')
 const getRandomPort = require('get-port')
-const hostname = process.env.HOSTNAME || '127.0.0.1'
-
-const svg = require('./lib/serve-svg')
-// const getPNG = require('./lib/get-png')
-
-app.get('/:icon.svg', svg)
-// app.get('/:icon.png', getPNG)
+const host = process.env.HOSTNAME || '127.0.0.1'
 
 Promise.resolve(process.env.PORT || getRandomPort())
-  .then(port => {
-    return new Promise((resolve, reject) => {
-      app.listen(port, hostname, function(error) {
-        if (error) {
-          reject(error)
-        } else {
-          // this is the server
-          resolve(this)
-        }
-      })
-    })
-  })
+  .then(port => createServer({host, port}))
   .then(server => {
     const {address, port} = server.address()
     console.warn('listening on %s',
